MSPSDS-927: Fixed stack too deep error and removed teams from shared and cosmetics
