--- cosmetics-web/app/models/team.rb
@@ -1,2 +1,0 @@
-class Team < Shared::Web::Team
-end
