--- cosmetics-web/app/models/team_user.rb
@@ -1,2 +1,0 @@
-class TeamUser < Shared::Web::TeamUser
-end
