+++ shared-web/app/models/shared/web/user.rb
@@ -5,18 +5,9 @@ module Shared
 
       belongs_to :organisation
 
-      has_many :team_users, dependent: :nullify
-      has_many :teams, through: :team_users
-
       field :first_name
       field :last_name
       field :email
-
-      def teams
-        # has_many through seems not to work with ActiveHash
-        # It's not well documented but the same fix has been suggested here: https://github.com/zilkey/active_hash/issues/25
-        team_users.map(&:team)
-      end
 
       def self.find_or_create(user)
         User.find_by(id: user[:id]) || User.create(user)
